de533be5 COMP: remove extraneous cast to double

+++ Modules/Filtering/ImageNoise/include/itkNoiseBaseImageFilter.hxx
@@ -62,7 +62,7 @@ NoiseBaseImageFilter<TInputImage, TOutputImage>
     }
   else if(value <= (double)NumericTraits<OutputImagePixelType>::NonpositiveMin() )
     {
-    return (double)NumericTraits<OutputImagePixelType>::NonpositiveMin();
+    return NumericTraits<OutputImagePixelType>::NonpositiveMin();
     }
   else if (NumericTraits<OutputImagePixelType>::is_integer)
     {
